Added ThankYou.md
Contributed to this Welcome Wall project by adding to the README.md a paragraph at the bottom to thank @bonn0062 for the warm welcome to newbies.  Additionally, I (hopefully) linked to and added the ThankYou.md file to the repository for other newbies to easily contribute by adding their thank you messages and images too.